% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_attgt.R
\name{covid_attgt}
\alias{covid_attgt}
\title{covid_attgt}
\usage{
covid_attgt(g, tp, data, xformla, ...)
}
\description{
takes groups and time periods and turns it into
an estimate of a group time average treatment effect
and a corresponding influence function
}
