% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_attgt.R
\name{matching_attgt}
\alias{matching_attgt}
\title{matching_attgt}
\usage{
matching_attgt(
  gt_data,
  xformla,
  matching_formula = xformla,
  ret_imputations = TRUE,
  ...
)
}
\arguments{
\item{gt_data}{data that is "local" to a particular group-time average
treatment effect}

\item{xformla}{one-sided formula for covariates used in the propensity score
and outcome regression models}

\item{matching_formula}{if provided, allows for the formula used to construct
the matches to be different from the outcome regression}

\item{ret_imputations}{whether or not to return actual and imputed outcomes
for the treated group}

\item{...}{extra function arguments; not used here}
}
\value{
attgt_if
}
\description{
Takes a "local" data.frame and computes
an estimate of a group time average treatment effect
and a corresponding influence function.

The code relies on \code{this.data} having certain variables defined.
In particular, there should be an \code{id} column (individual identifier),
\code{G} (group identifier), \code{period} (time period), \code{name}
(equal to "pre" for pre-treatment periods and equal to "post" for post
treatment periods), \code{Y} (outcome).

In our case, we call \code{pte::two_by_two_subset} which sets up the
data to have this format before the call to \code{matching_attgt}.
}
